refactor: add policy content to folder
