+++ app/shared/components/blocks/privacy-policy/intro-section/IntroSection.tsx
@@ -3,9 +3,9 @@ import { Box, Typography } from '@mui/material';
 import { theme } from '~/ds-components/theme/Theme';
 
 import { styles } from './IntroSection.styles';
-import { TipTapDoc } from '~/types/types/common.types';
+import { type TipTapDoc } from '~/types/types/common.types';
 
-import { PolicyContent } from '~/shared/components/blocks/privacy-policy/PolicyContent';
+import { PolicyContent } from '~/shared/components/blocks/privacy-policy/policy-content/PolicyContent';
 
 type IntroSectionProps = {
   title: string;
