+++ app/shared/components/blocks/privacy-policy/policy-content/PolicyContent.tsx
@@ -1,9 +1,9 @@
-import { SxProps, Theme } from '@mui/material';
+import type { SxProps, Theme } from '@mui/material';
 import Typography from '@mui/material/Typography';
 
 import TipTapContent from '~/components/tip-tap-content/TipTapContent';
 
-import { TipTapDoc } from '~/types/types/common.types';
+import type { TipTapDoc } from '~/types/types/common.types';
 
 export function PolicyContent({ doc, paragraphSx }: { doc: TipTapDoc; paragraphSx?: SxProps<Theme> }) {
   if (!doc) return null;
